fetchGithub to TS
