+++ backend/src/helpers/projectManagement.ts
@@ -1,5 +1,10 @@
 const mongo = require('../../src/database/DbServices');
-const userMng = require('./userManagement')
+import { jiraDecryptPassword } from './userManagement';
+const emptyScenario = require('../../src/models/emptyScenario');
+const emptyBackground = require('../../src/models/emptyBackground');
+import { writeFile } from '../../src/serverHelper';
+import { XMLHttpRequest } from 'xmlhttprequest';
+
 enum Sources {
     GITHUB = "github",
     JIRA = "jira",
@@ -31,7 +36,7 @@ class Repository {
 async function getJiraRepos(jiraUser: any) {
     if(!jiraUser)return []
     let { Host, AccountName, Password, Password_Nonce, Password_Tag } = jiraUser;
-    const jiraClearPassword = userMng.jiraDecryptPassword(Password, Password_Nonce, Password_Tag);
+    const jiraClearPassword = jiraDecryptPassword(Password, Password_Nonce, Password_Tag);
     const repos = await requestJiraRepos(Host, AccountName, jiraClearPassword);
     return await storeJiraRepos(repos)
 }
@@ -116,15 +121,87 @@ function dbProjects(userId: string) {
 	});
 }
 
-function fetchGithubRepos() {
-
+function uniqueRepositories(repositories) {
+	const uniqueIds = [];
+	const unique = [];
+	for (const i in repositories) {
+		if (uniqueIds.indexOf(repositories[i]._id.toString()) <= -1) {
+			uniqueIds.push(repositories[i]._id.toString());
+			unique.push(repositories[i]);
+		}
+	}
+	return unique;
 }
 
-function fetchDbRepos() {
+function execRepositoryRequests(link, user, password, ownerId, githubId) {
+	return new Promise((resolve, reject) => {
+		const request = new XMLHttpRequest(); // use fetch
+		// get Issues from GitHub
+		request.open('GET', link, true, user, password);
+		request.send();
+		request.onreadystatechange = async () => {
+			if (request.readyState !== 4) return;
+			if (request.status !== 200) { reject(this.status); return; }
+			const data = JSON.parse(request.responseText);
+			const projects = [];
+			const gitReposFromDb = await mongo.getAllSourceReposFromDb('github');
+			let mongoRepo;
+			for (const repo of data) {
+				// if this Repository is not in the DB create one in DB
+				if (!gitReposFromDb.some((entry) => entry.repoName === repo.full_name)) {
+					mongoRepo = await mongo.createGitRepo(repo.owner.id, repo.full_name, githubId, ownerId);
+				} else {
+					mongoRepo = gitReposFromDb.find((element) => element.repoName === repo.full_name); // await mongo.getOneGitRepository(repo.full_name)
+					if (mongoRepo.gitOwner === githubId) mongo.updateOwnerInRepo(repo._id, ownerId, mongoRepo.owner);
+				}
+				const repoName = repo.full_name;
+				const proj = {
+					_id: mongoRepo._id,
+					value: repoName,
+					source: 'github'
+				};
+				projects.push(proj);
+			}
+			resolve(projects);
+		};
+	});
+}
 
+function ownRepositories(ownerId, githubId, githubName, token) {
+	if (!githubName && !token) return Promise.resolve([]);
+	return execRepositoryRequests('https://api.github.com/user/repos?per_page=100', githubName, token, ownerId, githubId);
+}
+
+function starredRepositories(ownerId, githubId, githubName, token) {
+	if (!githubName && !token) return Promise.resolve([]);
+	return execRepositoryRequests(`https://api.github.com/users/${githubName}/starred`, githubName, token, ownerId, githubId);
+}
+
+async function fuseStoryWithDb(story) {
+	const result = await mongo.getOneStory(parseInt(story.story_id, 10), story.storySource);
+	if (result !== null) {
+		story.scenarios = result.scenarios;
+		story.background = result.background;
+		story.lastTestPassed = result.lastTestPassed;
+	} else {
+		story.scenarios = [emptyScenario()];
+		story.background = emptyBackground();
+	}
+	story.story_id = parseInt(story.story_id, 10);
+	if (story.storySource !== 'jira') story.issue_number = parseInt(story.issue_number, 10);
+
+	const finalStory = await mongo.upsertEntry(story.story_id, story, story.storySource);
+	story._id = finalStory._id;
+	// Create & Update Feature Files
+	writeFile(story);
+	return story;
 }
 
 module.exports = {
     getJiraRepos,
-    dbProjects
+    dbProjects,
+	uniqueRepositories,
+	starredRepositories,
+	ownRepositories,
+	fuseStoryWithDb
 };
